requirements.txt
torch==1.7.1+cu110
torchvision==0.8.2+cu110
torchaudio==0.7.2
transformers==3.5.0
python-telegram-bot==13.11
pymongo==4.1.1